refactor: extract ActionExecutor from Agent (PR 2)
- Create browser_use/agent/action_executor.py with ActionExecutor class
- Move multi_act(), rerun_history(), log_completion() and 13 helpers
- Wire ActionExecutor into StepPipeline via property
- ActionPhase._execute_actions() uses pipeline.action_executor
- Add action_executor property to BaseStepPipeline ABC
- Keep backward-compatible wrappers on Agent
- Use lazy imports for method-specific dependencies
- Fix type hints and remove dead code from Agent
