+++ examples/getting_started/01_basic_search.py
@@ -1,3 +1,34 @@
+# """
+# Setup:
+# 1. Get your API key from https://cloud.browser-use.com/new-api-key
+# 2. Set environment variable: export BROWSER_USE_API_KEY="your-key"
+# """
+
+# import asyncio
+# import os
+# import sys
+
+# # Add the parent directory to the path so we can import browser_use
+# sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
+
+# from dotenv import load_dotenv
+
+# load_dotenv()
+
+# from browser_use import Agent, ChatBrowserUse
+
+
+# async def main():
+# 	llm = ChatBrowserUse(model='bu-2-0')
+# 	task = "Search Google for 'what is browser automation' and tell me the top 3 results"
+# 	agent = Agent(task=task, llm=llm)
+# 	await agent.run()
+
+
+# if __name__ == '__main__':
+# 	asyncio.run(main())
+
+ 
 """
 Setup:
 1. Get your API key from https://cloud.browser-use.com/new-api-key
@@ -15,11 +46,17 @@ from dotenv import load_dotenv
 
 load_dotenv()
 
-from browser_use import Agent, ChatBrowserUse
+from browser_use.llm.deepseek.chat import ChatDeepSeek
+from browser_use import Agent
 
+# DEEPSEEK_API_KEY = os.getenv('DEEPSEEK_API_KEY')
 
 async def main():
-	llm = ChatBrowserUse(model='bu-2-0')
+	llm = ChatDeepSeek(
+		model="deepseek-chat",
+    	api_key=os.getenv('DEEPSEEK_API_KEY'),
+    	base_url="https://api.deepseek.com"
+	)
 	task = "Search Google for 'what is browser automation' and tell me the top 3 results"
 	agent = Agent(task=task, llm=llm)
 	await agent.run()
@@ -27,3 +64,4 @@ async def main():
 
 if __name__ == '__main__':
 	asyncio.run(main())
+
